Add code examples
