+++ apps/docs/src/app/core/component-docs/calendar/calendar-docs.component.html
@@ -31,7 +31,7 @@
 <component-example>
     <fd-calendar-mobile-example></fd-calendar-mobile-example>
 </component-example>
-<code-example [exampleFiles]="calendarSingleSource"></code-example>
+<code-example [exampleFiles]="calendarMobileSource"></code-example>
 
 <separator></separator>
 
@@ -43,7 +43,7 @@
 <component-example>
     <fd-calendar-options-example></fd-calendar-options-example>
 </component-example>
-<code-example [exampleFiles]="calendarSingleSource"></code-example>
+<code-example [exampleFiles]="calendarOptionsSource"></code-example>
 
 <separator></separator>
 
@@ -55,7 +55,7 @@
 <component-example>
     <fd-calendar-grid-example></fd-calendar-grid-example>
 </component-example>
-<code-example [exampleFiles]="calendarSingleSource"></code-example>
+<code-example [exampleFiles]="calendarGridSource"></code-example>
 
 <separator></separator>
