+++ apps/docs/src/app/core/component-docs/calendar/calendar-docs.module.ts
@@ -13,7 +13,7 @@ import { CalendarProgrammaticallyChangeExampleComponent } from './examples/calen
 import { CalendarI18nExampleComponent } from './examples/calendar-i18n-example.component';
 import { CalendarI18nMomentExampleComponent } from './examples/calendar--i18n-moment-example.component';
 import { SegmentedButtonModule, CalendarModule, FundamentalNgxCoreModule } from '@fundamental-ngx/core';
-import { CalendarMobileExampleComponent } from './examples/calendar-mobile-example.component';
+import { CalendarMobileExampleComponent } from './examples/calendar-mobile-example/calendar-mobile-example.component';
 import { CalendarOptionsExampleComponent } from './examples/calendar-options-example/calendar-options-example.component';
 import { CalendarGridExampleComponent } from './examples/calendar-grid-example/calendar-grid-example.component';
