+++ apps/docs/src/app/core/component-docs/calendar/examples/calendar-mobile-example/calendar-mobile-example.component.html
@@ -1,0 +1,45 @@
+<ng-template let-dialog let-dialogConfig="dialogConfig" #landScapeDialog>
+    <fd-dialog [dialogConfig]="dialogConfig" [dialogRef]="dialog">
+        <fd-dialog-body>
+            <fd-calendar #landscapeCalendar [mobileLandscape]="true" [ngModel]="datePicked" (closeClicked)="dialog.close()"></fd-calendar>
+        </fd-dialog-body>
+
+        <fd-dialog-footer>
+            <fd-dialog-footer-button>
+                <button fd-button
+                        fdType="emphasized"
+                        fd-dialog-decisive-button
+                        (click)="dialog.close('Continue'); dateChanged(landscapeCalendar.selectedDate)">
+                    Ok
+                </button>
+            </fd-dialog-footer-button>
+        </fd-dialog-footer>
+    </fd-dialog>
+</ng-template>
+
+<button fd-button (click)="openLandScapeDialog(landScapeDialog)">Open Calendar in landscape mobile mode</button>
+
+<ng-template let-dialog let-dialogConfig="dialogConfig" #portraitDialog>
+    <fd-dialog [dialogConfig]="dialogConfig" [dialogRef]="dialog">
+        <fd-dialog-body>
+            <fd-calendar #portraitCalendar [mobilePortrait]="true" [ngModel]="datePicked" (closeClicked)="dialog.close()"></fd-calendar>
+        </fd-dialog-body>
+
+        <fd-dialog-footer>
+            <fd-dialog-footer-button>
+                <button fd-button
+                        fdType="emphasized"
+                        fd-dialog-decisive-button
+                        (click)="dialog.close('Continue'); dateChanged(portraitCalendar.selectedDate);">
+                    Ok
+                </button>
+            </fd-dialog-footer-button>
+        </fd-dialog-footer>
+    </fd-dialog>
+</ng-template>
+
+<button fd-button (click)="openPortraitDialog(portraitDialog)">Open Calendar in portrait mobile mode</button>
+
+<div>
+    <label fd-label>Date Picked: {{ datePicked.toDateString() }}</label>
+</div>
